requirements.txt
molo.core==3.7.3
molo.profiles==1.2.6
psycopg2==2.6.2
django-modelcluster>=2.0,<3.0
setuptools==25.1.6
